+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/
 spec/v2.0.0.html).
 
+## [0.0.16] - 2026-03-30
+### Changed
+- Updated the low-speed vehicle type query to avoid NHTSA 404s.
+
 ## [0.0.15] - 2026-03-30
 ### Changed
 - Included upstream metadata and request IDs in `/cars/random` error responses for easier debugging.